Move renderer.ts into webgpu folder.
